refactor: Refactor authentication and authorization
Changing signatures and decoupling authorization and authentication
