+++ pkg/s3-proxy/authx/authentication/basic-auth.go
@@ -1,23 +1,28 @@
-package middlewares
+package authentication
 
 import (
 	"fmt"
 	"net/http"
 
+	"github.com/oxyno-zeta/s3-proxy/pkg/s3-proxy/authx/models"
 	"github.com/oxyno-zeta/s3-proxy/pkg/s3-proxy/config"
+	"github.com/oxyno-zeta/s3-proxy/pkg/s3-proxy/server/middlewares"
 	"github.com/oxyno-zeta/s3-proxy/pkg/s3-proxy/server/utils"
 	"github.com/thoas/go-funk"
+	"golang.org/x/net/context"
 )
+
+var bucketRequestContextKey = &contextKey{name: "bucket-request-context"}
 
 // nolint:whitespace
 func basicAuthMiddleware(basicConfig *config.BasicAuthConfig,
 	basicAuthUserConfigList []*config.BasicAuthUserConfig, templateConfig *config.TemplateConfig) func(http.Handler) http.Handler {
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			logEntry := GetLogEntry(r)
+			logEntry := middlewares.GetLogEntry(r)
 			path := r.URL.RequestURI()
 			// Get bucket request context from request
-			brctx := GetBucketRequestContext(r)
+			brctx := middlewares.GetBucketRequestContext(r)
 
 			// Get basic auth information
 			username, password, ok := r.BasicAuth()
@@ -62,6 +67,16 @@ func basicAuthMiddleware(basicConfig *config.BasicAuthConfig,
 				}
 				return
 			}
+
+			// Create Basic auth user
+			buser := &models.BasicAuthUser{Username: username}
+
+			// Add user to request context by creating a new context
+			ctx := context.WithValue(r.Context(), userContextKey, buser)
+			// Create new request with new context
+			r = r.WithContext(ctx)
+
+			logEntry.Info("Basic auth user %s authenticated", username)
 
 			next.ServeHTTP(w, r)
 		})
