+++ pkg/s3-proxy/authx/authentication/main_test.go
@@ -1,6 +1,6 @@
 // +build unit
 
-package middlewares
+package authentication
 
 import (
 	"reflect"
