+++ pkg/s3-proxy/authx/models/basic-auth-user.go
@@ -1,0 +1,11 @@
+package models
+
+const BasicAuthUserType = "BASIC"
+
+type BasicAuthUser struct {
+	Username string
+}
+
+func (u *BasicAuthUser) GetType() string {
+	return BasicAuthUserType
+}
