+++ pkg/s3-proxy/authx/models/generic-user.go
@@ -1,0 +1,7 @@
+package models
+
+// Generic user interface used to know specific type of user
+type GenericUser interface {
+	// Get type of user
+	GetType() string
+}
