+++ pkg/s3-proxy/authx/models/oidc-user.go
@@ -1,0 +1,12 @@
+package models
+
+const OIDCUserType = "OIDC"
+
+type OIDCUser struct {
+	Email  string
+	Groups []string
+}
+
+func (u *OIDCUser) GetType() string {
+	return OIDCUserType
+}
